chore: separate form and creation
