+++ front/src/features/session-definition/DefinitionForm.less
