+++ front/src/features/session-definition/style.less
@@ -1,2 +1,2 @@
 @import '../../styles/mixins';
-@import './Create';
+@import './DefinitionForm';
